Fix data loss on startup by replaying WAL after loading indexes
- Added WAL replay logic to `AletheiaDB::with_unified_config` to recover data written after the last persisted snapshot.
- Extracted `replay_wal` logic from `CheckpointManager` to `src/storage/recovery.rs` for reuse.
- Fixed `segment_reader` to handle zero-filled pre-allocated regions as EOF instead of corruption.
- Fixed test isolation issues in `tests/index_persistence.rs` where tests shared the default WAL directory.
- Added regression test `tests/regression_wal_replay.rs`.
